requirements.txt
ccxt
python-telegram-bot==13.15
pandas
python-dotenv
requests
